--- src/test/java/OOPSExamples/ClassA.java
@@ -1,19 +1,0 @@
-package OOPSExamples;
-
-public class ClassA {
-
-	private int a =10;
-
-	public int getA() {
-		return a;
-	}
-
-	public void setA(int a) {
-		this.a = a;
-	}
-
-	public void display()
-	{
-		System.out.println("Value in ClassA :- "+ a);
-	}
-}
