+++ src/test/java/oopsExamples/ClassB.java
@@ -1,4 +1,4 @@
-package OOPSExamples;
+package oopsExamples;
 
 public class ClassB extends ClassA
 {
